scripts/requirements.txt
boto3==1.34.0
psycopg2-binary==2.9.7
pandas==2.1.4
openpyxl==3.1.2